BACKPORT: UPSTREAM: smccc: add ARM_SMCCC_OWNER_STANDARD_HYP define
This provides a mechanism for querying which paravirtualized time
features are available in this hypervisor.

Also add the header file which defines the ABI for the paravirtualized
time features we're about to add.

(cherry picked from commit b48c1a45a190898103cec28771efc399fd65a05a)

BUG=b:154568514
TEST=various tests on rpi4 and kukui-arc-r

+++ virt/kvm/arm/psci.c
@@ -401,7 +401,7 @@ static int kvm_psci_call(struct kvm_vcpu *vcpu)
 int kvm_hvc_call_handler(struct kvm_vcpu *vcpu)
 {
 	u32 func_id = smccc_get_function(vcpu);
-	u32 val = SMCCC_RET_NOT_SUPPORTED;
+	long val = SMCCC_RET_NOT_SUPPORTED;
 	u32 feature;
 
 	switch (func_id) {
